Hi Haddock: Lex and rename Haddock docstrings in GHC, storing the results in interface files
This patch also restricts the limitation of `-haddock` being incompatible with
`Opt_KeepRawTokenStream`.

Names appearing in Haddock docstrings are lexed and renamed like any other names
appearing in the AST. We currently rename names irrespective of the namespace,
so both type and constructor names corresponding to an identifier will appear in
the docstring. Haddock will select a given name as the link destination based on
its own heuristics.

The export and documenation structure is now computed in GHC and serialised in
.hi files. This can be used by haddock to directly generate doc pages without
reparsing or renaming the source.

Hi Haddock: Lex and rename Haddock docstrings in GHC, storing the results in
interface files

This patch also restricts the limitation of `-haddock` being incompatible with
`Opt_KeepRawTokenStream`.

Names appearing in Haddock docstrings are lexed and renamed like any other names
appearing in the AST. We currently rename names irrespective of the namespace,
so both type and constructor names corresponding to an identifier will appear in
the docstring. Haddock will select a given name as the link destination based on
its own heuristics.

The export and documenation structure is now computed in GHC and serialised in
.hi files. This can be used by haddock to directly generate doc pages without
reparsing or renaming the source.