Clarify laws of TestEquality
It is unclear what `TestEquality` is for. There are 3 possible choices.

Assuming

```haskell
data Tag a where
    TagInt1 :: Tag Int
    TagInt2 :: Tag Int
```

## Weakest -- type param equality only

`Just Refl` merely means the type params are equal, the values being
compared might not be

```haskell
instance TestEquality where
    testEquality TagInt1 TagInt1 = Just Refl
    testEquality TagInt1 TagInt2 = Just Refl
    testEquality TagInt2 TagInt1 = Just Refl
    testEquality TagInt2 TagInt2 = Just Refl
```

## Middle -- Like `Eq`

`Just Refl` means the type params are equal, and the values are equal
according to `Eq`.

```haskell
instance TestEquality where
    testEquality TagInt1 TagInt1 = Just Refl
    testEquality TagInt2 TagInt2 = Just Refl
    testEquality _ _ = Nothing
```

## Strongest -- unique value concrete type

`Just Refl` means the type params are equal, and the values are equal,
and the class assume if the type params are equal the values must also
be equal. In other words, the type is a singleton type when the type
parameter is a closed term.

```haskell
-- instance TestEquality -- invalid instance because two variants for `Int`
```

------

The discussion in
https://github.com/haskell/core-libraries-committee/issues/21 on this
topic is veering towards the weakest option so that is what is
implementing here, so that is what I am implementing. The committee will
vote on this change.