change `Backend` type and remove direct dependencies
With this change, `Backend` is no longer a sum type with exposed value
constructors.  Decisions that were formerly made by asking "is the
current back end equal to (or different from) this named value
constructor?" are now made by interrogating the back end about
its properties.  The properties and the functions that interrogate
them are defined in `GHC.Driver.Backend.Types`.

The `Backend` type is currently a product (record) type, but it could
be make a sum type with only localized changes.  For an idea of what
that would look like, see `GHC.Driver.Backend.Legacy`.